starting to refactor the code such that variables are evaluated during expression evaluation rather then parsing. Now just need to refactor parsing to make this work
