+++ app/src/main/kotlin/org/example/expression/Expression.kt
@@ -1,0 +1,90 @@
+package org.example.expression
+import org.example.mainParse
+import org.typemeta.funcj.control.Either
+import org.typemeta.funcj.data.Chr
+import org.typemeta.funcj.parser.Parser
+import org.typemeta.funcj.parser.Text.dble
+import org.typemeta.funcj.parser.Text.ws
+import java.util.*
+
+
+sealed class Expression {
+
+    class Calculation( val op: Operator) : Expression()
+    {
+        lateinit var operands:Array<out Expression>
+        override val numOperands: Int
+            get() {
+                return operands.size
+            }
+
+        constructor(op: Operator, vararg operands : Expression) : this(op) {
+            if (operands.size == 2 && operands[1].numOperands == 2) {
+                var newOps:Array<Expression> = Array(2) { n -> N(n) }
+                var a = operands[0]
+                var b = operands[1]
+                if (a is N && b is Calculation && b.op.level == this.op.level) {
+                    var b1 = b.operands[0]
+                    var b2 = b.operands[1]
+                    newOps[0] = Calculation(this.op, a, b1)
+                    newOps[1] = b2
+                    this.operands = newOps
+                }
+                else {
+                    this.operands = operands
+                }
+            }
+            else this.operands = operands
+        }
+        override fun evaluate(context:Context): Either<String, N> {
+            return op.evaluate(context, *operands)
+        }
+
+        override fun toString():String {
+            var retVal = "("
+            for (i in operands) {
+                retVal += "$i $op "
+            }
+            if (operands.size > 0) {
+                retVal = retVal.substring(0, (retVal.length-op.toString().length - 2))
+            }
+            retVal += ")"
+            return retVal
+        }
+
+    }
+
+
+
+    abstract val numOperands: Int
+    abstract fun evaluate(context: Context): Either<String, N>
+
+
+    companion object {
+        var variables: MutableMap<String, N> = Hashtable()
+        val d = dble.andL(ws.many())!!
+        val du = d.and(U.parser.andL(ws.many()).optional())!!
+        val duExpr = du.map { a, b -> if (b.isEmpty) N(a) else N.NWithUnit(a, b.get()) }!!
+        fun lookup(name:String):Optional<N> = Optional.ofNullable(variables[name])
+        fun addVariable(name:String, value:N) {
+            variables.put(name, value)
+        }
+        fun dbleExpr() : Parser<Chr, N> {
+            return ws.many().andR(duExpr.or(vLookup()))
+        }
+        fun vLookup() : Parser<Chr, N> {
+            val v = mainParse.Companion.stringChoice(variables.keys.toList()).andL(ws.many())
+            return v.map{a -> lookup(a).get()}!!
+        }
+        fun resetVars() {
+            variables.clear()
+        }
+
+
+    }
+
+
+}
+
+
+
